Move some models and constants around
